requirements.txt
aiogram>=3.4.0
aiosqlite>=0.20.0
Pillow>=10.2.0
pytest>=8.0.0
